refactor resource recovery time constants
