+++ src/features/game/events/goldMine.ts
@@ -1,5 +1,6 @@
 import Decimal from "decimal.js-light";
 import cloneDeep from "lodash.clonedeep";
+import { GOLD_RECOVERY_TIME } from "../lib/constants";
 import { GameState, Rock } from "../types/game";
 
 export type GoldMineAction = {
@@ -12,9 +13,6 @@ type Options = {
   action: GoldMineAction;
   createdAt?: number;
 };
-
-// 24 hours
-export const GOLD_RECOVERY_TIME = 24 * 60 * 60;
 
 export enum MINE_ERRORS {
   NO_PICKAXES = "No pickaxes left",
