+++ src/features/game/expansion/components/resources/Gold.tsx
@@ -13,6 +13,7 @@ import pickaxe from "assets/tools/iron_pickaxe.png";
 import {
   GRID_WIDTH_PX,
   GOLD_MINE_STAMINA_COST,
+  GOLD_RECOVERY_TIME,
 } from "features/game/lib/constants";
 import { Context } from "features/game/GameProvider";
 import { ToastContext } from "features/game/toast/ToastQueueProvider";
@@ -23,10 +24,7 @@ import { getTimeLeft } from "lib/utils/time";
 import { miningAudio, miningFallAudio } from "lib/utils/sfx";
 import { HealthBar } from "components/ui/HealthBar";
 import { LandExpansionRock } from "features/game/types/game";
-import {
-  GOLD_RECOVERY_TIME,
-  EVENT_ERRORS,
-} from "features/game/events/landExpansion/mineGold";
+import { EVENT_ERRORS } from "features/game/events/landExpansion/mineGold";
 import { calculateBumpkinStamina } from "features/game/events/landExpansion/replenishStamina";
 import { Overlay } from "react-bootstrap";
 import { Label } from "components/ui/Label";
