+++ src/features/game/expansion/components/resources/Iron.tsx
@@ -13,6 +13,7 @@ import pickaxe from "assets/tools/stone_pickaxe.png";
 import {
   GRID_WIDTH_PX,
   IRON_MINE_STAMINA_COST,
+  IRON_RECOVERY_TIME,
 } from "features/game/lib/constants";
 import { Context } from "features/game/GameProvider";
 import { ToastContext } from "features/game/toast/ToastQueueProvider";
@@ -23,10 +24,7 @@ import { getTimeLeft } from "lib/utils/time";
 import { miningAudio, miningFallAudio } from "lib/utils/sfx";
 import { HealthBar } from "components/ui/HealthBar";
 import { LandExpansionRock } from "features/game/types/game";
-import {
-  IRON_RECOVERY_TIME,
-  MINE_ERRORS,
-} from "features/game/events/landExpansion/ironMine";
+import { MINE_ERRORS } from "features/game/events/landExpansion/ironMine";
 import { calculateBumpkinStamina } from "features/game/events/landExpansion/replenishStamina";
 import { Overlay } from "react-bootstrap";
 import { Label } from "components/ui/Label";
